### 2025-08-27 1f4c2eb
Shrink workflow and use case examples
