+++ joss/paper.md
@@ -74,28 +74,17 @@ Existing metagenomic methods lack the statistical rigor to control false positiv
 
 To address these challenges, YACHT offers a statistical framework that can accurately determine the presence or absence of microbial genome in a sample through hypothesis testing. The algorithm’s mathematical model accounts for evolutionary sequence divergence and incomplete sequencing depth by utilizing genome similarity and minimum sequencing depth parameters. It employs the FrachMinHash sketching technique, an alignment-free $k$-mer approach, facilitating fast and accurate genome detection that can efficiently process large datasets. YACHT ensures precise detection of low abundance taxa with a user-defined false negative rate, minimizing the risk of misinterpretation of the result. Our approach can be used for other metagenomic applications such as functional profiling, metatranscriptomic studies [@marcelino2019metatranscriptomics], metabolic potential analyses [@ward2018metapoap; @pereira2024metatranscriptomics], and the characterization of low abundant clinical metagenomic samples such as skin [@godlewska2020metagenomic]. YACHT enhances metagenomic analysis by offering reduced reliance on arbitrary thresholds, improving the interpretability of the result without compromising biological relevance, and allowing researchers to differentiate between genuine artifacts from “noise” with statistical confidence.
 
-![the workflow of the YACHT algorithm for deciding inclusion of microbial genome in samples. The process involves four main steps: Generate k-mer sketches for both reference genomes and samples using sourmash; Preprocess the reference genomes using the yacht train command to merge redundant genomes based on ANI thresholds, retaining the most representative genome; Run the YACHT algorithm to perform probabilistic hypothesis testing with user-defined parameters.](workflow.png)
-
 # Workflow
 
-## 1. Create sketches
+The YACHT workflow involves four primary steps. First, `yacht sketch` samples compact representations of reference genomes using `sourmash`. Second, `yacht train` preprocesses the reference genomes, merging those with high average nucleotide identity (ANI) into a single representative. Third, `yacht run` executes the core YACHT algorithm to perform hypothesis testing and determine the presence or absence of organisms. Finally, `yacht convert` transforms the results into popular output formats like CAMI, BIOM, and GraphPhlAn.
 
-This step uses the command `yacht sketch`, which generates sketches for both the samples and the reference genomes. Users must utilize sourmash to extract sketches from a reference database of microbial genomes. [sourmash Databases](https://sourmash.readthedocs.io/en/latest/databases.html) provide a variety of pre-formed databases of such sketches, or users can create a custom database using the `sourmash sketch` command on FASTA/FASTQ files of reference genomes (see the [sourmash documentation](https://sourmash.readthedocs.io/en/latest/)). Other available databases include the [GTDB genomic representatives database](https://gtdb.ecogenomic.org/downloads). The sketches for samples must be generated using the same $k$-mer size and scale factor as those used for the reference database. The scale factor acts as an indicator of data compression, with smaller values being more appropriate for smaller datasets.
 
-## 2. Preprocess the reference genomes
+**For the detailed step-by-step workflow, see [the repository](https://github.com/KoslickiLab/YACHT/).
 
-This step uses the command `yacht train`, which identifies and merges genomes that are roughly identical based on Average Nucleotide Identity (ANI). It extracts sketches from the reference database and converts them into a binary format specific to YACHT. For any two organisms with ANI greater than a user-specified threshold, one will be removed from the reference set, as such organisms are considered too similar to be "distinguishable."
+![](workflow.png)
 
[…]
