### 2025-08-27 92e19d7
Add DOI to multiple articles in paper.bib
