+++ joss/paper.bib
@@ -3,6 +3,7 @@
   author={Tian, Shuchang and Kim, Min Soo and Zhao, Jingcheng and Heber, Kerim and Hao, Fuhua and Koslicki, David and Tian, Sangshan and Singh, Vishal and Patterson, Andrew D and Bisanz, Jordan E},
   journal={Cell Host \& Microbe},
   year={2025},
+  doi={10.1016/j.chom.2025.02.007},
   publisher={Elsevier}
 }
 @article{hera2023deriving,
@@ -13,6 +14,7 @@
   number={7},
   pages={1061--1068},
   year={2023},
+  doi={10.1101/gr.277651.123},
   publisher={Cold Spring Harbor Lab}
 }
 @article{blanca2022statistics,
@@ -23,6 +25,7 @@
   number={2},
   pages={155--168},
   year={2022},
+  doi={10.1089/cmb.2021.0431},
   publisher={Mary Ann Liebert, Inc., publishers 140 Huguenot Street, 3rd Floor New~…}
 }
 @article{irber2024sourmash,
@@ -32,7 +35,8 @@
   volume={9},
   number={98},
   pages={6830},
-  year={2024}
+  year={2024},
+  doi={10.21105/joss.06830}
 }
 
 @article{koslicki2024yacht,
@@ -43,6 +47,7 @@
   number={2},
   pages={btae047},
   year={2024},
+  doi={10.1093/bioinformatics/btae047},
   publisher={Oxford University Press}
 }
 
@@ -54,6 +59,7 @@
   number={24},
   pages={4284--4286},
   year={2018},
+  doi={10.1093/bioinformatics/bty510},
   publisher={Oxford University Press}
 }
 
@@ -65,6 +71,7 @@
   volume={10},
   pages={1--10},
   year={2019},
+  doi={10.1186/s43008-019-0012-8},
   publisher={Springer}
 }
 
@@ -76,6 +83,7 @@
   number={1},
   pages={2323235},
   year={2024},
+  doi={10.1080/19490976.2024.2323235},
   publisher={Taylor \& Francis}
 }
 
@@ -86,6 +94,7 @@
   volume={77},
   pages={3201--3212},
   year={2020},
+  doi={10.1007/s00284-020-02163-4}
   publisher={Springer}
 }
 
@@ -97,6 +106,7 @@
   number={6},
   pages={669--681},
   year={2012},
+  doi={10.1093/bib/bbs054},
   publisher={Oxford University Press}
 }
 
@@ -108,6 +118,7 @@
   number={6},
   pages={1882--1899},
   year={2013},
+  doi={10.1111/1462-2920.12086}
   publisher={Wiley Online Library}
 }
 
@@ -119,6 +130,7 @@
   number={11},
   pages={1063--1071},
   year={2017},
+  doi={10.1038/nmeth.4458},
   publisher={Nature Publishing Group US New York}
 }
 
@@ -130,6 +142,7 @@
   number={4},
   pages={429--440},
   year={2022},
+  doi={10.1038/s41592-022-01431-4},
   publisher={Nature Publishing Group US New York}
 }
 
@@ -142,6 +155,7 @@
   number={5},
   pages={e01424--22},
   year={2022},
+  doi={10.1128/mbio.01424-22},
   publisher={Am Soc Microbiol}
 }
 
@@ -153,6 +167,7 @@
   number={1},
   pages={707},
   year={2023},
+  doi={10.1038/s41597-023-02622-0},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -163,6 +178,7 @@
   volume={9},
   pages={848136},
   year={2022},
+  doi={10.3389/fmars.2022.848136},
   publisher={Frontiers Media SA}
 }
 
[…]
