### 2025-08-27 dd713ba
low abundance update
