### 2025-08-27 70f8efa
Update paper.bib
Added missing comma was added that was causing the bib file test fail.
