+++ joss/paper.bib
@@ -94,7 +94,7 @@
   volume={77},
   pages={3201--3212},
   year={2020},
-  doi={10.1007/s00284-020-02163-4}
+  doi={10.1007/s00284-020-02163-4},
   publisher={Springer}
 }
 
@@ -118,7 +118,7 @@
   number={6},
   pages={1882--1899},
   year={2013},
-  doi={10.1111/1462-2920.12086}
+  doi={10.1111/1462-2920.12086},
   publisher={Wiley Online Library}
 }
 
