### 2025-08-28 6ebedc8
Update runner to use ubuntu-latest
For PR #33
