+++ .github/workflows/runTest.yml
@@ -2,7 +2,7 @@ name: runTest
 on: [push]
 jobs:
   test-python-app:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     defaults:
       run:
         shell: bash -el {0}
