### 2025-08-28 2e089e5
fixed some security issues detected by SonarCloud
