### 2025-08-28 2b0b827
fixed some security issues detected by SonarCloud
