### 2025-08-28 0737a1d
fixed some security issues detected by SonarCloud
