### 2025-08-28 39b0570
fixed some security issues detected by SonarCloud
