+++ .github/workflows/runTest.yml
@@ -43,7 +43,7 @@ jobs:
         run: pytest tests/ --cov-report term-missing --cov-report xml:tests.xml --cov=yacht
     
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v4.1.0
+        uses: codecov/codecov-action@eaaf4bedf32543fbe8914606212be542791f6c28 # v4.1.0
         with:
           files: ./tests.xml
         env:
