### 2025-08-28 188cd55
fixed some security issues detected by SonarCloud
