### 2025-09-23 #9066 @editorialbot
[REVIEW]: YACHT: an ANI-based statistical test to detect microbial presence/absence in a metagenomic sample
**Submitting author:** @akp6031 (<a href="http://orcid.org/0009-0003-4104-2960">Adam Park</a>)
**Repository:** https://github.com/KoslickiLab/YACHT/
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v1.3.1
