### 2025-12-03 4d2b279
update meta.yaml for bioconda submission
