+++ conda_recipe/meta.yaml
@@ -1,7 +1,8 @@
+{% set name = "YACHT" %}
 {% set version = "1.3.1" %}
 
 package:
-  name: yacht
+  name: {{ name|lower }}
   version: {{ version }}
 
 source:
@@ -9,11 +10,16 @@ source:
   sha256: 66864bfd12292f78ca7ba91899e17701e1dfeffb95478665154dc7fc16248248
 
 build:
-  number: 0
-  skip: True  # [osx]
-  script: "{{ PYTHON }} -m pip install . --no-deps --no-build-isolation --no-cache-dir -vvv"
+  number: 1
+  script:
+    - sed -i.bak 's|find_packages|find_namespace_packages|' setup.py
+    - sed -i.bak 's|g++|$(CXX)|' Makefile
+    - sed -i.bak 's|-lpthread|-pthread|' Makefile
+    - sed -i.bak 's|-Wsign-compare|-Wsign-compare -I$(PREFIX)/include|' Makefile
+    - rm -rf *.bak
+    - {{ PYTHON }} -m pip install . --no-deps --no-build-isolation --no-cache-dir --use-pep517 -vvv
   run_exports:
-    - {{ pin_subpackage('yacht') }}
+    - {{ pin_subpackage('yacht', max_pin="x") }}
 
 requirements:
   build:
@@ -31,7 +37,6 @@ requirements:
     - tqdm
     - biom-format
     - numpy >=1.22.4
-    - setuptools
     - requests
 
   run:
@@ -59,18 +64,18 @@ test:
     - yacht --help
 
 about:
-  home: https://github.com/KoslickiLab/YACHT
-  summary: 'YACHT is a mathematically rigorous hypothesis test for the presence or absence of organisms in a metagenomic sample, based on average nucleotide identity (ANI).'
+  home: "https://github.com/KoslickiLab/YACHT"
+  summary: "YACHT is a mathematically rigorous hypothesis test for the presence or absence of organisms in a metagenomic sample, based on average nucleotide identity (ANI)."
   license: MIT
   license_family: MIT
-  license_file: LICENSE.txt
-  dev_url: https://github.com/KoslickiLab/YACHT
-  doc_url: https://github.com/KoslickiLab/YACHT/wiki
+  license_file: "LICENSE.txt"
+  dev_url: "https://github.com/KoslickiLab/YACHT"
+  doc_url: "https://github.com/KoslickiLab/YACHT/wiki"
 
 extra:
-  skip-lints:
-    - should_use_compilers
-    - should_be_noarch_generic
+  additional-platforms:
+    - linux-aarch64
+    - osx-arm64
   identifiers:
     - doi:10.1093/bioinformatics/btae047
   recipe-maintainers:
