### 2025-12-03 6bc34f9
update sha256 hash in the meta.yaml
