+++ conda_recipe/meta.yaml
@@ -7,7 +7,7 @@ package:
 
 source:
   url: https://github.com/KoslickiLab/YACHT/releases/download/v{{ version  }}/yacht-{{ version  }}.tar.gz
-  sha256: 66864bfd12292f78ca7ba91899e17701e1dfeffb95478665154dc7fc16248248
+  sha256: e0a3d729b7c71b3f13ded2c6d6970a7aaf3f436baa04865fc17ce01300c8ba4e
 
 build:
   number: 1
