### 2025-12-03 866c048
update meta.yaml file
