+++ conda_recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "YACHT" %}
-{% set version = "1.3.1" %}
+{% set version = "1.3.2" %}
 
 package:
   name: {{ name|lower }}
@@ -7,10 +7,10 @@ package:
 
 source:
   url: https://github.com/KoslickiLab/YACHT/releases/download/v{{ version  }}/yacht-{{ version  }}.tar.gz
-  sha256: e0a3d729b7c71b3f13ded2c6d6970a7aaf3f436baa04865fc17ce01300c8ba4e
+  sha256: 8ff83a130c51923d867fc3277257c1566b24a94069bb77da39e8783529dbe4b1
 
 build:
-  number: 1
+  number: 0
   script:
     - sed -i.bak 's|find_packages|find_namespace_packages|' setup.py
     - sed -i.bak 's|g++|$(CXX)|' Makefile
