### 2025-12-05 00e2db0
First commit on superyacht branch, including files required to calculate effective coverage, etc according to sylph (Shaw and Yu, 2024).
