+++ hypothesis_recovery_src.py
@@ -1,0 +1,480 @@
+import os
+import sys
+import numpy as np
+import warnings
+from scipy.stats import binom
+from scipy.special import betaincinv
+import pandas as pd
+import zipfile
+from tqdm import tqdm
+from multiprocessing import Pool
+import sourmash
+import glob
+from typing import List, Set, Tuple
+from .utils import load_signature_with_ksize, decompress_all_sig_files
+# Configure Loguru logger
+from loguru import logger
+from cov_calc import cov_calc
+
+warnings.filterwarnings("ignore")
+
+
+\
+logger.remove()
+logger.add(
+    sys.stdout, format="{time:YYYY-MM-DD HH:mm:ss} - {level} - {message}", level="INFO"
+)
+
+SIG_SUFFIX = ".sig"
+
+
+def get_organisms_with_nonzero_overlap(
+    manifest: pd.DataFrame,
+    sample_file: str,
+    scale: int,
+    ksize: int,
+    num_threads: int,
+    path_to_genome_temp_dir: str,
+    path_to_sample_temp_dir: str,
+) -> List[str]:
+    """
+    This function runs the sourmash multisearch to find the organisms that have non-zero overlap with the sample.
+    :param manifest: a dataframe with the following columns:
+                        'organism_name',
+                        'md5sum',
+                        'num_unique_kmers_in_genome_sketch',
+                        'num_total_kmers_in_genome_sketch',
+                        'genome_scale_factor',
+                        'num_exclusive_kmers_in_sample_sketch',
+                        'num_total_kmers_in_sample_sketch',
+                        'sample_scale_factor',
+                        'min_coverage'
+    :param sample_file: string (path to the sample signature file)
+    :param scale: int (scale factor)
+    :param ksize: string (size of kmer)
+    :param num_threads: int (number of threads to use for parallelization)
+    :param path_to_genome_temp_dir: string (path to the genome temporary directory generated by the training step)
+    :param path_to_sample_temp_dir: string (path to the sample temporary directory)
+    :return: a list of organism names that have non-zero overlap with the sample
+    """
+    # run the sourmash multisearch
+    # prepare the input files for the sourmash multisearch
+    # unzip the sourmash signature file to the temporary directory
+    logger.info("Unzipping the sample signature zip file")
+    with zipfile.ZipFile(sample_file, "r") as sample_zip_file:
+        sample_zip_file.extractall(path_to_sample_temp_dir)
+    all_gz_files = glob.glob(f"{path_to_sample_temp_dir}/signatures/*.sig.gz")
+    # decompress all signature files
+    logger.info(f"Decompressing {len(all_gz_files)} .sig.gz files using {num_threads} threads.")
+    decompress_all_sig_files(all_gz_files, num_threads)
+
+    sample_sig_file = pd.DataFrame(
+        [
+            os.path.join(path_to_sample_temp_dir, "signatures", sig_file)
+            for sig_file in os.listdir(
+                os.path.join(path_to_sample_temp_dir, "signatures")
+            )
+        ]
+    )
+    sample_sig_file_path = os.path.join(path_to_sample_temp_dir, "sample_sig_file.txt")
+    sample_sig_file.to_csv(sample_sig_file_path, header=False, index=False)
+
+    organism_sig_file = pd.DataFrame(
+        [
+            os.path.join(path_to_genome_temp_dir, "signatures", md5sum + SIG_SUFFIX)
+            for md5sum in manifest["md5sum"]
+        ]
+    )
+    organism_sig_file_path = os.path.join(
+        path_to_sample_temp_dir, "organism_sig_file.txt"
+    )
+    organism_sig_file.to_csv(organism_sig_file_path, header=False, index=False)
+
+    # run the sourmash multisearch
[…]
