+++ utils.py
@@ -1,860 +1,0 @@
-import os
-import sys
-import numpy as np
-from typing import List, Optional
-from collections import Counter
-from sourmash import load_one_signature
-from collections import defaultdict
-from tqdm import tqdm
-import pandas as pd
-from multiprocessing import Pool
-from loguru import logger
-from typing import Optional, List, Set, Dict, Tuple, Any
-import shutil
-import gzip
-import math
-import random
-import sourmash
-from dataclasses import dataclass
-from glob import glob
-
-# Configure Loguru logger
-logger.remove()
-logger.add(
-    sys.stdout, format="{time:YYYY-MM-DD HH:mm:ss} - {level} - {message}", level="INFO"
-)
-
-# Set up constants
-COL_NOT_FOUND_ERROR = "Column not found: {}"
-FILE_LOCATION = os.path.dirname(os.path.realpath(__file__))
-# Adding two more contstants (RTR)
-SAMPLE_SIZE_CUTOFF: int = 25 
-PVALUE_CUTOFF: float = 0.9999999999
-ksize = 31 #Note: hard-coding this for now
-
-# Set up global variables
-__version__ = "2.0.1"
-GITHUB_API_URL = "https://api.github.com/repos/KoslickiLab/YACHT/contents/demo/{path}"
-GITHUB_RAW_URL = "https://raw.githubusercontent.com/KoslickiLab/YACHT/main/demo/{path}"
-BASE_URL = "https://farm.cse.ucdavis.edu/~ctbrown/sourmash-db/"
-ZENODO_COMMUNITY_URL = "https://zenodo.org/api/records/?communities=yacht&size=100"
-
-# A dataclass to implement something equivalent to sylph's rust-based enum implementation (AdjustStatus)
-@dataclass(frozen=True)
-class AdjustStatusLambda:
-    value: float
-
-class AdjustStatusHigh:
-    pass
-
-class AdjustStatusLow:
-    pass
-
-class AdjustStatusNone:
-    pass
-
-# Class for cov_calc output
-@dataclass
-class AniResult:
-    naive_ani: float
-    final_est_ani: float
-    final_est_cov: float
-    seq_name: str
-    gn_name: str
-    contig_name: str
-    mean_cov: float
-    median_cov: float
-    containment_index: Tuple[int, int]
-    lambda_status: AdjustStatusLambda
-    ani_ci: Tuple[Optional[float], Optional[float]]
-    lambda_ci: Tuple[Optional[float], Optional[float]]
-    genome_sketch: Any
-    rel_abund: Optional[float]
-    seq_abund: Optional[float]
-    kmers_lost: Optional[int]
-
-# Class for cov_calc arguments
-# Note that these are being set here for now, but could be brought as CLI arguments for yacht
-class _ContainArgs:
-    def __init__(self):
-        self.ci_int = True
-        self.ratio = True
-        self.mme = True
-        self.nb = True
-        self.mle = True
-        self.min_count_correct = 1 # Example value
-
-def load_signature_with_ksize(filename: str, ksize: int) -> sourmash.SourmashSignature:
-    """
-    Helper function that loads the signature for a given kmer size from the provided signature file.
-    Filename should point to a sourmash signature file. Raises exception if given kmer size is not present in the file.
-    This is a wrapper of sourmash.load_file_as_signatures, and accept all types of format: .sig, .sig.zip, .sbt, .lca, and .sqldb.
-    However, this function specifically ask for 1 signature so lca format is not appropriate; as of sourmash v4.8, sqldb doesn't accept "abund" parameter for signatures.
-    :param filename: string (location of the signature file of any format: .sig, .sig.zip, .sbt, .lca, and .sqldb)
-    :param ksize: int (size of kmer)
[…]
