+++ src/yacht/hypothesis_recovery_src.py
@@ -14,11 +14,12 @@ from typing import List, Set, Tuple
 from .utils import load_signature_with_ksize, decompress_all_sig_files
 # Configure Loguru logger
 from loguru import logger
+from cov_calc import cov_calc
 
 warnings.filterwarnings("ignore")
 
 
-
+\
 logger.remove()
 logger.add(
     sys.stdout, format="{time:YYYY-MM-DD HH:mm:ss} - {level} - {message}", level="INFO"
@@ -143,6 +144,8 @@ def get_exclusive_hashes(
             2. the number of unique hashes exclusive to the organism under consideration that are in the sample
         a new manifest dataframe that only contains the organisms that have non-zero overlap with the sample
     """
+    pvalue_cutoff=0.9999999999
+    min_count_thresh=3 #TODO: consider whether to change this value
 
     def __find_exclusive_hashes(
         md5sum: str,
@@ -155,7 +158,7 @@ def get_exclusive_hashes(
             os.path.join(path_to_temp_dir, "signatures", md5sum + SIG_SUFFIX), ksize
         )
         return {hash for hash in sig.minhash.hashes if hash in single_occurrence_hashes}
-
+    
     # get manifest information for the organisms that have non-zero overlap with the sample
     sub_manifest = manifest.loc[
         manifest["organism_name"].isin(nontrivial_organism_names), :
@@ -177,6 +180,10 @@ def get_exclusive_hashes(
                 multiple_occurrence_hashes.add(hash)
             else:
                 single_occurrence_hashes.add(hash)
+
+
+    #print(multiple_occurrence_hashes)
+    
     del multiple_occurrence_hashes  # free up memory
 
     # Find hashes that are unique to each organism
@@ -188,10 +195,32 @@ def get_exclusive_hashes(
                 md5sum, path_to_genome_temp_dir, ksize, single_occurrence_hashes
             )
         )
+
+    #print(f"Single occurrence hashes")
+    #print(single_occurrence_hashes) #adding this for testing
     del single_occurrence_hashes  # free up memory
 
     # Get sample hashes
     sample_hashes = set(sample_sig.minhash.hashes)
+   
+     # Get sample hashes keys 
+    sample_hashes_keys = sample_sig.minhash.hashes.keys()
+    #print(sample_hashes_keys)
+    samp_kmers_items = sample_sig.minhash.hashes.items()
+    samp_dict = dict(samp_kmers_items)
+    
+    stats_list = []
+    for md5sum in tqdm(organism_md5sum_list, desc="Processing coverage per organism"):
+            sig = load_signature_with_ksize(
+                os.path.join(path_to_genome_temp_dir, "signatures", md5sum + SIG_SUFFIX),
+                ksize,
+            )
+            stats_out = cov_calc(sample_sig, sig) #location of cov_calc, which calculates effective coverage and other things according to Shaw and Yu (2024)
+            stats_list.append(stats_out)
+    
+    final_stats_df = pd.concat(stats_list, ignore_index=True)
+    
+    del stats_list # free up memory
 
     # Find hashes that are unique to each organism and in the sample
     logger.info("Finding hashes that are unique to each organism and in the sample")
@@ -199,12 +228,41 @@ def get_exclusive_hashes(
     for i, exclusive_hashes in enumerate(
         tqdm(exclusive_hashes_org, desc="Matching exclusive hashes with sample")
     ):
+        #print(f"exclusive_hash", exclusive_hashes)
         exclusive_hashes_info.append(
             (len(exclusive_hashes), len(exclusive_hashes.intersection(sample_hashes)))
         )
 
-    return exclusive_hashes_info, sub_manifest
+    # Calculate lambda and other related coverage metrics for each organism in the sample
[…]

[…]
