+++ src/yacht/cov_calc.py
@@ -54,9 +54,7 @@ def cov_calc(sample_sig: sourmash.SourmashSignature, genome_sig: sourmash.Sourma
     covs = []
     contain_count = 0
     for kmer in gn_hashes:
-        #print(kmer)
         if kmer in sample_hashes_keys:
-            #print(f"Overlap")
             if samp_dict[kmer] == 0:
                 continue
             contain_count += 1
@@ -69,26 +67,18 @@ def cov_calc(sample_sig: sourmash.SourmashSignature, genome_sig: sourmash.Sourma
         1/ksize)
 
     covs.sort()
-    #print(covs)
 
     if len(covs) == 0: 
-        #print("Zero length")
         covs.append(0) 
     
-
-    #cov_set = set(covs)
     len_ind = len(covs)//2
-    #print("len_ind")
-    #print(len_ind)
     median_cov = covs[len(covs)//2]
-    #print(median_cov)
+
     pois_obj = poisson(median_cov) #creates a discrete frozen Poisson distribution object
     cov_max = float('inf')
     
     if median_cov < 30: #if median coverage of 30 is not fulfilled
-        #print(f"Below 30")
         for i in range(len_ind,len(covs), 1):
-            #print(i)
             cov = covs[i]
             if pois_obj.cdf(cov) < PVALUE_CUTOFF:
                 cov_max = cov
@@ -101,7 +91,6 @@ def cov_calc(sample_sig: sourmash.SourmashSignature, genome_sig: sourmash.Sourma
         if cov <= cov_max:
             full_covs.append(cov)
     var = variation(full_covs)
-    #print("Variation is:", var)
     if var is not None:
         logger.debug("VAR {} {}", var, genome_sig.name)
 
@@ -109,7 +98,6 @@ def cov_calc(sample_sig: sourmash.SourmashSignature, genome_sig: sourmash.Sourma
     geq1_mean_cov = sum(full_covs)//len(covs)    
     if median_cov > MEDIAN_ANI_THRESHOLD:
         return_lambda = ADJUST_STATUS_HIGH
-        #print(f"Above_threshold: {type(return_lambda).__name__}") #for testing
 
     else:
         if (myArgs.ratio == True):
@@ -128,53 +116,38 @@ def cov_calc(sample_sig: sourmash.SourmashSignature, genome_sig: sourmash.Sourma
         else:
             return_lambda = AdjustStatusLambda(value=test_lambda) # Wrap the float in the dataclass
             
-    #print(f"Return lambda type: {type(return_lambda).__name__}")
-    
     match return_lambda:
 
         case AdjustStatusLambda(value=lam):
-            #print(f"Case1")
             # executes if it is the Lambda case
             final_est_cov = lam
             opt_lambda = final_est_cov
-            #print(f"Status is Lambda, coverage set to: {final_est_cov:.2f}")
 
         case AdjustStatusHigh():
             # executes if it is high coverage case
-            #print(f"Case2")
             if median_cov < MAX_MEDIAN_FOR_MEAN_FINAL_EST:
                 final_est_cov = geq1_mean_cov
-                #print(f"Status is High, using geq1_mean_cov logic")
             else:
                 final_est_cov = median_cov
-                #print(f"Status is High, using median_cov logic")
             opt_lambda = final_est_cov
 
         case AdjustStatusLow():
-            #print(f"Case3")
             # if it is the "low" case
             # final_est_cov logic is handled elsewhere, or use a default
             opt_lambda = None
-            #print("Status is Low, using naive_ani logic later")
 
         # Adding a "wild-card" case, just to be safe
         case _:
[…]
