### 2025-12-10 74beed8
add txt ouput format
