### 2025-12-10 36b6829
added Irber LC, Brooks PT, Reiter TE, Pierce-Ward NT, Hera MR, Koslicki D, Brown CT. 2022. Lightweight compositional analysis of metagenomes with FracMinHash and minimum metagenome covers. bioRxiv 10.1101/2022.01.11.475838 [DOI] per JOSS reviewer request
