+++ joss/paper.bib
@@ -330,3 +330,12 @@
   doi={10.1093/femsre/fuad012},
   publisher={Oxford University Press}
 }
+
+@article{Irber2022FracMinHash,
+  author    = {Irber, Luiz C. and Brooks, Patrick T. and Reiter, Travis E. and Pierce-Ward, Nathan T. and Hera, M. R. and Koslicki, David and Brown, C. Titus},
+  title     = {Lightweight compositional analysis of metagenomes with FracMinHash and minimum metagenome covers},
+  journal   = {bioRxiv},
+  year      = {2022},
+  doi       = {10.1101/2022.01.11.475838},
+  url       = {https://doi.org/10.1101/2022.01.11.475838}
+}
