### 2025-12-11 9be54d7
addressing review comment 137.4 and aboffin review comment on use case example section of the main JOSS paper
