### 2025-12-11 5ee940f
updated bib with Singer, Esther, et al. "Next generation sequencing data of a defined microbial mock community." Scientific data 3.1 (2016): 1-8.
