+++ joss/paper.bib
@@ -339,3 +339,14 @@
   doi       = {10.1101/2022.01.11.475838},
   url       = {https://doi.org/10.1101/2022.01.11.475838}
 }
+
+@article{Singer2016MockCommunity,
+  author    = {Singer, Esther and Bushnell, Brian and Coleman-Derr, Devin and Douglas, Gina M. and Bowman, Benjamin and Bowers, Robert M. and Levy, Adi and Gies, Esther and Cheng, Jan-Fang and Copeland, Alex and others},
+  title     = {Next generation sequencing data of a defined microbial mock community},
+  journal   = {Scientific Data},
+  volume    = {3},
+  number    = {1},
+  pages     = {1--8},
+  year      = {2016},
+  publisher = {Nature Publishing Group}
+}
