### 2025-12-11 3154e64
updated table 1 caption with data used and yacht parameters used to generate results
