### 2025-12-12 f258a86
Update title to include 'Software' in paper.md
Change title by adding "Software for" (refer to JOSS reviewer #137)
