### 2025-12-12 4e7fe9f
Update summary of YACHT tool in paper.md
added description of what statistical framework is being referred to
