### 2025-12-17 93b9582
Update workflow to trigger on pull requests
