+++ .github/workflows/runTest.yml
@@ -1,5 +1,8 @@
 name: runTest
 on: [push, pull_request]
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 jobs:
   test-python-app:
     runs-on: ubuntu-latest
