### 2025-12-22 89e853f
Response to JOSS reviews
readme - Added target audiences and the problem statement at the front, briefly wrote the dependencies.
JOSS/paper.md - figure formatting, references from and within figures and tables, input clarifications.
