### 2025-12-30 50ff040
Various code improvements to improve legibility and remove stray bits that aren't necessary.
