### 2026-01-02 4dd7972
Various bug fixes and improvements to code quality. Replaced relevant print statements with logger. Moved all constants to utils.py.
