### 2026-01-06 92e25cc
Changed to a more typically python enum for cov_calc.
