+++ joss/paper.bib
@@ -350,3 +350,9 @@
   year      = {2016},
   publisher = {Nature Publishing Group}
 }
+@book{irber2020decentralizing,
+  title={Decentralizing indices for genomic data},
+  author={Irber Jr, Luiz Carlos},
+  year={2020},
+  publisher={University of California, Davis}
+}
