### 2026-01-08 ea10ad8
Fix citation formatting and enhance YACHT description
Corrected citation formatting on markdown (main body). It should be a semicolon, not a comma when listing multiple citations.
