### 2026-01-10 80bedab
Minor update to README and .gitignore.
