### 2026-01-11 #9066 @csoneson
Thank you @aboffin!
Regarding authorship, this appears to be in agreement with JOSS's [authorship policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship).
Could you please tick off the remaining boxes in your checklist (also the one related to performance)?

@Vini2, @anuradhawick - when you have a moment, please have a look at the responses provided by the authors in the respective GitHub issues and let us know what you think. Thanks all!
