### 2026-01-12 b1bd726
Added winner map functionality to hypothesis_recovery_src.py.
