+++ src/yacht/hypothesis_recovery_src.py
@@ -10,11 +10,18 @@ from tqdm import tqdm
 from multiprocessing import Pool
 import sourmash
 import glob
-from typing import List, Set, Tuple
-from .utils import load_signature_with_ksize, decompress_all_sig_files
+from typing import List, Set, Tuple, Dict
+from .utils import load_signature_with_ksize, decompress_all_sig_files, MIN_ANI_THRESHOLD
 # Configure Loguru logger
 from loguru import logger
 from .cov_calc import cov_calc
+
+"""
+Hypothesis Recovery and Coverage Analysis Module
+
+This module implements YACHT's core statistical framework for organism detection
+in metagenomic samples, with integrated coverage modeling and abundance estimation.
+"""
 
 warnings.filterwarnings("ignore")
 
@@ -193,12 +200,8 @@ def get_exclusive_hashes(
 
     # Get sample hashes
     sample_hashes = set(sample_sig.minhash.hashes)
-   
-     # Get sample hashes keys 
-    sample_hashes_keys = sample_sig.minhash.hashes.keys()
-    samp_kmers_items = sample_sig.minhash.hashes.items()
-    samp_dict = dict(samp_kmers_items)
-    
+
+    # Calculate coverage statistics for each organism
     stats_list = []
     for md5sum in tqdm(organism_md5sum_list, desc="Processing coverage per organism"):
             sig = load_signature_with_ksize(
@@ -207,9 +210,12 @@ def get_exclusive_hashes(
             )
             stats_out = cov_calc(sample_sig, sig) #location of cov_calc, which calculates effective coverage and other things according to Shaw and Yu (2024)
             stats_list.append(stats_out)
-    
+
     final_stats_df = pd.concat(stats_list, ignore_index=True)
-    
+
+    # Add organism_name to final_stats_df for merging (stats are in same order as sub_manifest)
+    final_stats_df['organism_name'] = sub_manifest['organism_name'].values
+
     del stats_list # free up memory
 
     # Find hashes that are unique to each organism and in the sample
@@ -222,27 +228,121 @@ def get_exclusive_hashes(
             (len(exclusive_hashes), len(exclusive_hashes.intersection(sample_hashes)))
         )
 
-    # Calculate lambda and other related coverage metrics for each organism in the sample
-    #logger.info("Calculate lambda for each organism in the sample")
-    #for i, lambda_stats in enumerate()
-
-    columns_of_interest = [
-    'naive_ani', 
-    'final_est_ani', 
-    'final_est_cov', 
-    'mean_cov', 
-    'median_cov',
-    'lambda_ci',
-    'ani_ci'
-    ]
-
-    # Select only those columns from the DataFrame
-    selected_data = final_stats_df[columns_of_interest]
-    summary_stats = selected_data.describe()
-
-    #print(summary_stats)
+    logger.info("Building winner map for k-mer reassignment and relative abundance estimation")
+    winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize)
+    final_stats_df = estimate_relative_abundance(final_stats_df, winner_map, sample_sig)
 
     return exclusive_hashes_info, sub_manifest, final_stats_df
+
+
+def build_winner_map(
+    final_stats_df: pd.DataFrame,
+    path_to_genome_temp_dir: str,
+    ksize: int
+) -> Dict[int, Tuple[float, str]]:
+    """
+    Creates a "winner map" procedure that assigns k-mers to the organism with the highest ANI.
+
+    This implements the "winner takes all" strategy from sylph (Shaw and Yu, 2024) where
+    shared k-mers are assigned to the organism with the best ANI match, preventing double-counting.
+    Please note that this differs from the approach in sylph in that the procedure is run once, rather than
+    twice.
+
[…]
