### 2026-01-14 1f81fcc
Added parallelization to the coverage calculation; should increase speed.
