+++ src/yacht/hypothesis_recovery_src.py
@@ -119,13 +119,33 @@ def get_organisms_with_nonzero_overlap(
     return multisearch_result["match_name"].to_list()
 
 
+def _calculate_coverage_worker(args):
+    """
+    Worker function for parallel coverage calculation.
+
+    :param args: Tuple of (md5sum, path_to_genome_temp_dir, ksize, sample_sig)
+    :return: Result from cov_calc or None if error occurs
+    """
+    md5sum, path_to_genome_temp_dir, ksize, sample_sig = args
+    try:
+        sig = load_signature_with_ksize(
+            os.path.join(path_to_genome_temp_dir, "signatures", md5sum + SIG_SUFFIX),
+            ksize,
+        )
+        return cov_calc(sample_sig, sig)
+    except Exception as e:
+        logger.warning(f"Error calculating coverage for {md5sum}: {e}")
+        return None
+
+
 def get_exclusive_hashes(
     manifest: pd.DataFrame,
     nontrivial_organism_names: List[str],
     sample_sig: sourmash.SourmashSignature,
     ksize: int,
     path_to_genome_temp_dir: str,
-) -> Tuple[List[Tuple[int, int]], pd.DataFrame]:
+    num_threads: int = 16,
+) -> Tuple[List[Tuple[int, int]], pd.DataFrame, pd.DataFrame]:
     """
     This function gets the unique hashes exclusive to each of the organisms that have non-zero overlap with the sample, and
     then find how many are in the sample.
@@ -143,11 +163,13 @@ def get_exclusive_hashes(
     :param sample_sig: the sample signature
     :param ksize: int (size of kmer)
     :param path_to_genome_temp_dir: string (path to the genome temporary directory generated by the training step)
+    :param num_threads: int (number of threads to use for parallel coverage calculation, default: 16)
     :return:
         a list of tuples, each tuple contains the following information:
             1. the number of unique hashes exclusive to the organism under consideration
             2. the number of unique hashes exclusive to the organism under consideration that are in the sample
         a new manifest dataframe that only contains the organisms that have non-zero overlap with the sample
+        a dataframe with coverage statistics for each organism
     """
     def __find_exclusive_hashes(
         md5sum: str,
@@ -201,15 +223,28 @@ def get_exclusive_hashes(
     # Get sample hashes
     sample_hashes = set(sample_sig.minhash.hashes)
 
-    # Calculate coverage statistics for each organism
-    stats_list = []
-    for md5sum in tqdm(organism_md5sum_list, desc="Processing coverage per organism"):
-            sig = load_signature_with_ksize(
-                os.path.join(path_to_genome_temp_dir, "signatures", md5sum + SIG_SUFFIX),
-                ksize,
+    # Calculate coverage statistics for each organism (parallelized)
+    logger.info(f"Calculating coverage statistics using {num_threads} threads")
+    with Pool(processes=num_threads) as pool:
+        # Prepare arguments for parallel processing
+        args_list = [
+            (md5sum, path_to_genome_temp_dir, ksize, sample_sig)
+            for md5sum in organism_md5sum_list
+        ]
+        # Use imap for progress tracking with tqdm
+        stats_list = list(
+            tqdm(
+                pool.imap(_calculate_coverage_worker, args_list),
+                total=len(organism_md5sum_list),
+                desc="Processing coverage per organism"
             )
-            stats_out = cov_calc(sample_sig, sig) #location of cov_calc, which calculates effective coverage and other things according to Shaw and Yu (2024)
-            stats_list.append(stats_out)
+        )
+
+    # Filter out None results (from errors)
+    stats_list = [stats for stats in stats_list if stats is not None]
+
+    if not stats_list:
[…]
