+++ README.md
@@ -390,59 +390,4 @@ yacht convert --yacht_output 'result.xlsx' --sheet_name 'min_coverage0.01' --gen
 
 </br>
 
-## Methods
-
-YACHT integrates three complementary approaches for organism detection and quantification:
-
-### 1. Hypothesis Testing (Presence/Absence)
-- Uses **exclusive k-mers** (unique to each organism) for statistical testing
-- Binomial test determines if enough exclusive k-mers observed to reject null hypothesis
-- Accounts for mutation rate and minimum coverage thresholds
-- Output: `in_sample_est` (True/False), `p_vals`
-
-### 2. Coverage Modeling (ANI & Abundance)
-- Implements Shaw & Yu (2024) method from [sylph](https://github.com/bluenote-1577/sylph)
-- Estimates effective ANI and coverage accounting for sequencing depth variation
-- Calculates expected coverage (lambda) using Poisson distribution modeling
-- More accurate than naive containment-based ANI
-- Output: `final_est_ani`, `final_est_cov`, coverage statistics
-
-### 3. Winner Map K-mer Reassignment (Relative Abundance)
-- Follows "winner takes all" strategy from sylph
-- Shared k-mers between organisms assigned to organism with highest ANI
-- Prevents double-counting in relative abundance calculations
-- Single-pass design: uses coverage results without recalculation
-- Output: `rel_abund` (normalized), `kmers_lost`
-
-### Workflow
-```
-Sample Input
-    ↓
-Filter organisms (any k-mer overlap)
-    ↓
-Find exclusive k-mers (for hypothesis testing)
-    ↓
-Calculate coverage & ANI (Shaw & Yu method)
-    ↓
-Build winner_map (k-mer reassignment)
-    ↓
-Estimate relative abundance
-    ↓
-Run hypothesis tests
-    ↓
-Merge results + Filter by ANI threshold
-    ↓
-Excel Output
-```
-
-### Key Features
-- **Exclusive k-mers**: Prevent false positives from shared sequences
-- **Coverage adjustment**: More accurate ANI estimates
-- **Single-pass performance**: Efficient computation
-- **ANI filtering**: Removes low-quality matches (< 90% ANI)
-
-### References
-1. Koslicki, D., White, S., Ma, C., & Novikov, A. (2024). YACHT: an ANI-based statistical test to detect microbial presence/absence in a metagenomic sample. *Bioinformatics*, 40(2), btae047.
-2. Shaw, J., & Yu, Y. W. (2024). Rapid species-level metagenome profiling and containment estimation with sylph. *Nature Biotechnology*. https://doi.org/10.1038/s41587-024-02412-y
-
 
