### 2026-01-14 94cb13a
Added sample_sig as a pool variable to remove big performance overheads.
