+++ src/yacht/hypothesis_recovery_src.py
@@ -119,7 +119,7 @@ def get_organisms_with_nonzero_overlap(
     return multisearch_result["match_name"].to_list()
 
 
-# Adding a global variable for sharing sample_sig across worker processes (reduces overhead)
+# Global variable for sharing sample signature across worker processes
 _worker_sample_sig = None
 
 def _init_coverage_worker(sample_sig):
@@ -134,7 +134,7 @@ def _init_coverage_worker(sample_sig):
 def _calculate_coverage_worker(args):
     """
     Worker function for parallel coverage calculation.
-    Uses global _worker_sample_sig instead of passing it as argument to avoid overhead.
+    Uses global _worker_sample_sig instead of passing it as argument to avoid pickling overhead.
 
     :param args: Tuple of (md5sum, path_to_genome_temp_dir, ksize)
     :return: Result from cov_calc or None if error occurs
@@ -158,6 +158,8 @@ def get_exclusive_hashes(
     ksize: int,
     path_to_genome_temp_dir: str,
     num_threads: int = 16,
+    winner_takes_all: bool = False,
+    batch_size: int = 1000,
 ) -> Tuple[List[Tuple[int, int]], pd.DataFrame, pd.DataFrame]:
     """
     This function gets the unique hashes exclusive to each of the organisms that have non-zero overlap with the sample, and
@@ -177,6 +179,8 @@ def get_exclusive_hashes(
     :param ksize: int (size of kmer)
     :param path_to_genome_temp_dir: string (path to the genome temporary directory generated by the training step)
     :param num_threads: int (number of threads to use for parallel coverage calculation, default: 16)
+    :param winner_takes_all: bool (enable winner-takes-all k-mer reassignment and relative abundance, default: False)
+    :param batch_size: int (batch size for winner-takes-all processing, default: 1000)
     :return:
         a list of tuples, each tuple contains the following information:
             1. the number of unique hashes exclusive to the organism under consideration
@@ -238,16 +242,21 @@ def get_exclusive_hashes(
 
     # Calculate coverage statistics for each organism (parallelized)
     logger.info(f"Calculating coverage statistics using {num_threads} threads")
+
+    # Calculate (organism) chunk size for progress visibility
+    chunk_size = max(1, len(organism_md5sum_list) // (num_threads * 50))
+    logger.info(f"Using chunk size of {chunk_size} for parallel processing")
+
     with Pool(processes=num_threads, initializer=_init_coverage_worker, initargs=(sample_sig,)) as pool:
-        # Prepare arguments for parallel processing (sample_sig shared to avoid overhead)
+        # Prepare arguments for parallel processing (sample_sig shared via initializer to avoid pickling overhead)
         args_list = [
             (md5sum, path_to_genome_temp_dir, ksize)
             for md5sum in organism_md5sum_list
         ]
-        # Use imap for progress tracking with tqdm
+        # Use imap_unordered for better performance (order doesn't matter for coverage stats)
         stats_list = list(
             tqdm(
-                pool.imap(_calculate_coverage_worker, args_list),
+                pool.imap_unordered(_calculate_coverage_worker, args_list, chunksize=chunk_size),
                 total=len(organism_md5sum_list),
                 desc="Processing coverage per organism"
             )
@@ -276,9 +285,25 @@ def get_exclusive_hashes(
             (len(exclusive_hashes), len(exclusive_hashes.intersection(sample_hashes)))
         )
 
-    logger.info("Building winner map for k-mer reassignment and relative abundance estimation")
-    winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize)
-    final_stats_df = estimate_relative_abundance(final_stats_df, winner_map, sample_sig)
[…]
