+++ src/yacht/run_YACHT.py
@@ -56,7 +56,7 @@ def add_arguments(parser):
     parser.add_argument(
         "--batch_size",
         type=int,
-        help="Batch size for winner-takes-all processing (lower size uses less memory)."
+        help="Batch size for winner-takes-all processing (lower size uses less memory). "
              "Only used with --winner_takes_all. Default: 1000",
         default=1000,
     )
