### 2026-01-15 6997b5c
Renamed MEDIAN_ANI_THRESHOLD to avoid confusion.
