### 2026-01-16 3333649
Fixed critical bug due to umap_unordered- now matching on organism name.
