### 2026-01-27 782b2fd
Added ANI-capping utility to avoid biologically impossible ANI inflation due to low lambda.
