### 2026-01-28 835f434
Updated cov_calc to cap ANIs at 1.0; pt II of ani capping.
