### 2026-01-29 8b36fb1
Removed old commented code that was replaced by the winner-map addition.
