+++ src/yacht/utils.py
@@ -32,7 +32,7 @@ FILE_LOCATION = os.path.dirname(os.path.realpath(__file__))
 SAMPLE_SIZE_CUTOFF: int = 25
 PVALUE_CUTOFF: float = 0.9999999999
 MIN_ANI_THRESHOLD: float = 0.90  # Minimum ANI threshold for filtering organisms
-MEDIAN_ANI_THRESHOLD: float = 2.00
+MEDIAN_ANI_THRESHOLD: float = 3.00
 MAX_MEDIAN_FOR_MEAN_FINAL_EST: float = 15.0
 MIN_COUNT_THRESH: int = 3
 LAMBDA_EPSILON: float = 1e-10  # Minimum lambda value to avoid dividing by zero
