### 2026-02-07 #9066 @csoneson
@akp6031 - the paper as it stands is quite a bit longer than stated in the JOSS [guidelines](https://joss.readthedocs.io/en/pre-2026-submission/paper.html#what-should-my-paper-contain) (250-1000 words). Could you look into shortening it a bit to get closer to the recommendations (looking at it, I think it should be possible to condense the text without reducing the information content)?
In addition, I think it would be helpful to readers to briefly clarify the relation to the original method paper - currently the summary describes YACHT (Koslicki et al., 2024) as a command-line tool, so it is not clear from here what the current submission brings (as you explained in the pre-review issue, there has been substantial development, and perhaps the reference to the earlier paper should rather focus on the methodology?).
There is also a missing DOI (see [post above](https://github.com/openjournals/joss-reviews/issues/9066#issuecomment-3863887178)), and it would be great if you could spell out 'USA' in the affiliations.

When this is done you can generate a new proof with `@editorialbot generate pdf`.
Thanks!
