### 2026-02-26 0c7f1f8
Revise summary to highlight YACHT's statistical framework
Updated the summary section to distinguish between YACHT methodology paper (Koslicki et al., 2024) and JOSS submission
