### 2026-03-02 8d1c109
word reduction, should be just under 1000
