+++ joss/paper.md
@@ -43,13 +43,13 @@ authors:
     corresponding: true 
     affiliation: "1, 2, 3"
 affiliations:
-  - name: School of Electrical Engineering and Computer Science, Pennsylvania State University, USA
+  - name: School of Electrical Engineering and Computer Science, Pennsylvania State University, United States of America
     index: 1
     ror: 04p491231
-  - name: Huck Institutes of the Life Sciences, Pennsylvania State University, USA
+  - name: Huck Institutes of the Life Sciences, Pennsylvania State University, United States of America
     index: 2
     ror: 04p491231
-  - name: Department of Biology, Pennsylvania State University, USA
+  - name: Department of Biology, Pennsylvania State University, United States of America
     index: 3
     ror: 04p491231
 date: 10 February 2025
@@ -58,25 +58,23 @@ bibliography: paper.bib
 
 # Summary
 
-In metagenomics, identifying genomes present in a sample is an important initial task, but is complicated by taxonomic profiling tools lacking uncertainty quantification and using incomplete reference databases missing exact genome matches. YACHT (**Y**es/No **A**nswers to **C**ommunity membership via **H**ypothesis **T**esting) [@koslicki2024yacht] introduces a statistical framework of taxonomic profiling that uses binomial hypothesis testing on exclusive k-mers to determine genome presence/absence in a metagenomic sample confidently. This paper describes the software implementation of YACHT, a command-line tool that translates the methodology into a practical, accessible tool for metagenomic analysis. YACHT assists in discovering rare microbiomes by identifying low-abundant species missed in other taxonomic profiling approaches while also controlling the false-negative rate. Its statistical model overcomes challenges in sequencing coverage and incomplete genomes, making it ideal for diverse metagenomic applications, including functional profiling, metatranscriptomics, and clinical microbiome analysis.
-
-YACHT presents a robust, $k$-mer sketching-based statistical framework for accurately detecting genetic similarity between the reference database and the metagenomic sample by incorporating evolutionary sequence divergence through the average nucleotide identity (ANI) and sequencing coverage to enable efficient detection of sampled genomes. The workflow for YACHT includes the following commands. To begin, `yacht sketch` creates reduced representation "sketches" of the reference and sample datasets enabling swift comparisons. Then, `yacht train` is used to find a representative of closely related reference genomes using ANI. Lastly, `yacht run` uses the YACHT algorithm to perform hypothesis testing and identify the presence or absence of species. YACHT is developed with C++ and Python and depends on `sourmash` [@irber2024sourmash], a program for extracting and managing $k$-mers.
+Identifying genomes in metagenomics samples can be complicated by taxonomic profiling tools that lack uncertainty quantification and rely on incomplete reference databases. YACHT (**Y**es/No **A**nswers to **C**ommunity membership via **H**ypothesis **T**esting) [@koslicki2024yacht] introduces a $k$-mer sketching based statistical framework that incorporates average nucleotide identity (ANI) and sequencing coverage to detect genetic similarity between reference and sample genomes using binomial hypothesis testing on exclusive $k$-mers to confidently determine genome presence/absence. This paper describes the software implementation of this methodology as a command-line tool that detects low-abundant species while controlling the false-negative rate, making it applicable to functional profiling, metatranscriptomics, and clinical microbiome analysis despite incomplete genomes and variable sequencing coverage. YACHT is developed with C++ and Python and depends on `sourmash` [@irber2024sourmash] for $k$-mer extraction and management.
 
 # Statement of need
 
[…]
