### 2026-03-12 d35725d
Patched a minor bug relating to performance of python's gzip on Mac OS, replacing it with the system gzip
