### 2026-03-13 #9066 @stephwon
Hi @csoneson,

Thank you for your feedback and apologies for the delayed follow-up. We've updated the manuscript with the following changes:

- Condensed the text to better align with the JOSS word count guidelines
- Clarified the relationship to the original method paper, focusing on the methodology rather than describing YACHT as a command-line tool
- Added the missing DOI
- Spelled out 'USA' in the affiliations

A new proof has been generated with `@editorialbot generate pdf`. Please let us know if any further changes are needed.
