### 2026-03-13 7f04e17
Added two-pass mode to --winner-take-all procedure.
