+++ src/yacht/hypothesis_recovery_src.py
@@ -11,7 +11,15 @@ from multiprocessing import Pool
 import sourmash
 import glob
 from typing import List, Set, Tuple, Dict
-from .utils import load_signature_with_ksize, decompress_all_sig_files, MIN_ANI_THRESHOLD
+from .utils import (
+    load_signature_with_ksize,
+    decompress_all_sig_files,
+    MIN_ANI_THRESHOLD,
+    ratio_lambda,
+    ani_from_lambda,
+    MIN_COUNT_THRESH,
+    SAMPLE_SIZE_CUTOFF,
+)
 # Configure Loguru logger
 from loguru import logger
 from .cov_calc import cov_calc
@@ -164,6 +172,7 @@ def get_exclusive_hashes(
     num_threads: int = 16,
     winner_takes_all: bool = False,
     batch_size: int = 1000,
+    two_pass: bool = True,
 ) -> Tuple[List[Tuple[int, int]], pd.DataFrame, pd.DataFrame]:
     """
     This function gets the unique hashes exclusive to each of the organisms that have non-zero overlap with the sample, and
@@ -185,6 +194,8 @@ def get_exclusive_hashes(
     :param num_threads: int (number of threads to use for parallel coverage calculation, default: 16)
     :param winner_takes_all: bool (enable winner-takes-all k-mer reassignment and relative abundance, default: False)
     :param batch_size: int (batch size for winner-takes-all processing, default: 1000)
+    :param two_pass: bool (use sylph's two-pass approach for more accurate k-mer reassignment, default: True)
+                     Set to False for original one-pass behavior (for testing/comparison)
     :return:
         a list of tuples, each tuple contains the following information:
             1. the number of unique hashes exclusive to the organism under consideration
@@ -291,8 +302,31 @@ def get_exclusive_hashes(
 
     # Conditionally run winner-takes-all (memory-intensive but provides relative abundance)
     if winner_takes_all:
-        logger.info("Building winner map for k-mer reassignment (winner-takes-all strategy)")
-        winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize, batch_size)
+        if two_pass:
+            # Two-pass approach (sylph-aligned): more accurate for closely related organisms
+            # Pass 1: Build initial winner map using original ANI estimates
+            logger.info("Pass 1: Building initial winner map for k-mer reassignment")
+            winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize, batch_size)
+
+            # Recalculate ANI using only won k-mers (prepares for Pass 2)
+            final_stats_df = recalculate_ani_from_winner_map(
+                final_stats_df, winner_map, sample_sig, ksize, batch_size
+            )
+
+            # Pass 2: Rebuild winner map with refined ANI estimates
+            # Only include organisms that weren't eliminated in the recalculation
+            logger.info("Pass 2: Rebuilding winner map with refined ANI estimates")
+            winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize, batch_size)
+        else:
+            # One-pass approach (original): faster but less accurate for related organisms
+            logger.info("One-pass mode: Building winner map with initial ANI estimates")
+            winner_map = build_winner_map(final_stats_df, path_to_genome_temp_dir, ksize, batch_size)
+
+            # Add placeholder columns for consistency
+            final_stats_df['reassignment_status'] = 'one_pass'
+            final_stats_df['original_ani'] = final_stats_df['final_est_ani'].copy()
+
+        # Calculate relative abundance using final winner map
         final_stats_df = estimate_relative_abundance(final_stats_df, winner_map, sample_sig, batch_size)
 
         # Free up memory by dropping genome_sketch column (no longer needed)
@@ -301,9 +335,11 @@ def get_exclusive_hashes(
[…]
