+++ joss/paper.md
@@ -64,7 +64,7 @@ Identifying genomes in metagenomics samples can be complicated by taxonomic prof
 
 Accurately identifying low-abundance microbial communities remains a significant challenge in metagenomics. Current methods rely on arbitrary filter thresholds that, even when applied, produce results skewed by sequencing errors and evolutionary processes, compromising profiling accuracy and leading to misinterpretations [@schloss2020removal; @jia2022sequencing]. The lack of a systematic credibility framework can undermine researcher confidence, a problem compounded by incomplete reference databases and variable sequencing coverage depth.
 
-Metagenomic methods depend on reference databases that are often incomplete and misaligned with taxonomic frameworks, leaving evolutionarily diverged microbes undetected and causing profiling inaccuracies [@kunin2008bioinformatician; @schlaberg2017validation; @loeffler2020improving; @marcelino2020ccmetagen]. Addressing this equires  analytical frameworks that incorporate genome similarity metrics, though sequencing coverage depth presents an additional challenge to reliable microbial detection.
+Metagenomic methods depend on reference databases that are often incomplete and misaligned with taxonomic frameworks, leaving evolutionarily diverged microbes undetected and causing profiling inaccuracies [@kunin2008bioinformatician; @schlaberg2017validation; @loeffler2020improving; @marcelino2020ccmetagen]. Addressing this requires analytical frameworks that incorporate genome similarity metrics, though sequencing coverage depth presents an additional challenge to reliable microbial detection.
 
 Sequence coverage depth—the portion of a microbe’s genome detected in a sample—is crucial for detecting low-abundance microbes, which are often misinterpreted as noise due to limited sequencing depth [@mande2012classification; @shakya2013comparative; @sczyrba2017critical; @meyer2022critical]. The lack of guidelines for biologically meaningful coverage depth thresholds introduces subjectivity, making dynamic coverage depth thresholds essential. Yet even with adequate coverage and reliable genome references, controlling statistical errors remains a major challenge.
 
@@ -111,7 +111,7 @@ We present the three use case examples to demonstrate the application of YACHT f
 
 **Low abundance samples:** YACHT can analyze metagenomic samples with low microbial DNA concentrations common in clinical and environmental studies. Using a human skin metagenomics samples, we show that ANI threshold and k-size markedly influence species specificity. See [Low abundance samples](https://github.com/KoslickiLab/YACHT/tree/main/use_case_examples/low_abundance_samples).
 
-**Metagenomic-assembled genome (MAG) fishing:** Using a single MAG as a training reference database, YACHT searches for specific MAGs within a sample. Applied to two skin metagenomic samples, results shows detection is sensitive to sequencing depth, coverage, and parameter choice. See [MAG fishing](https://github.com/KoslickiLab/YACHT/tree/main/use_case_examples/MAG_fishing).
+**Metagenomic-assembled genome (MAG) fishing:** Using a single MAG as a training reference database, YACHT searches for specific MAGs within a sample. Applied to two skin metagenomic samples, results shows that detection is sensitive to sequencing depth, coverage, and parameter choice. See [MAG fishing](https://github.com/KoslickiLab/YACHT/tree/main/use_case_examples/MAG_fishing).
 
 **Synthetic metagenomes:** YACHT verifies the presence of designed microbes in mock microbial communities. Higher ANI thresholds recover expected genomes while lower thresholds introduce false positives, demonstrating how ANI and minimum coverage parameters affect sensitivy and specificity. See [Synthetic metagenomes](https://github.com/KoslickiLab/YACHT/tree/main/use_case_examples/synthetic_metagenome)
 
