### 2026-03-15 #9066 @csoneson
Thanks @stephwon - I have gone through the paper and opened a [PR](https://github.com/KoslickiLab/YACHT/pull/147) with a few small edits (and one additional comment). Please have a look and merge if you agree; then we can move on to the next steps.
