### 2026-03-25 #9066 @stephwon
@csoneson We appreciate you bringing this to our attention. We have harmonized the terminology, reviewed the edits, and merged them. The `paper.md` file should reflect the latest edits and fixes. We look forward to hearing about the next steps for publication. Thank you!
