### 2026-03-28 #9066 @csoneson
Thanks @stephwon. I noticed two additional typos (sorry, missed them last time):
* on line 83: samples -> sample
* on line 91: sensitivy -> sensitivity
Could you fix those and generate a new proof with `@editorialbot generate pdf`?

Apart from that it looks good to me. I will make another post below with the next steps.
