### 2026-03-30 #9066 @stephwon
Hi @csoneson, thank you for catching those errors. We have fixed the issues and generated a [new proof](https://github.com/openjournals/joss-reviews/issues/9066#issuecomment-4151656125). We look forward to hearing about the next steps!
