### 2026-04-14 #9066 @stephwon
Hi @csoneson,
We had a slight issue with the Zenodo login access, which caused delays. We will provide an update on the checklist in a couple of days. Thank you for your patience!
