### 2026-04-15 #9066 @stephwon
Hi @csoneson,
Thank you for your patience, we looked over the checklist and can confirm the following:

- All authors on the manuscripts and their affiliations are matching
- YACHT version is v1.3.2, last updated on Dec 3, 2025 and no new update has been made since then.
- Have archived the release on [YACHT Zenodo](https://zenodo.org/records/19597912) and the associated DOI information is [https://doi.org/10.5281/zenodo.19597912](https://doi.org/10.5281/zenodo.19597912)
- The title, author list including ORCIDs in the archive match to the JOSS paper
- The license listed in the archive is MIT, which matches the software license in the GitHub repository (https://github.com/KoslickiLab/YACHT/blob/main/LICENSE.txt)
