### 2026-04-19 #9066 @csoneson
Thanks @stephwon. Could you please edit the metadata of the Zenodo record so that the displayed title, authors and license match the paper? At the moment I see the below for the title/author, and a Creative Commons license. Thanks!
