### 2026-04-19 #9066 @csoneson
Thanks @stephwon, this looks better 🙂 I think your first name is still missing in the Zenodo metadata

Could you also make the version tags the same in the GitHub repo (`v1.3.2`) and on Zenodo (`1.3.2`)?
