### 2026-04-19 #9066 @stephwon
Hi @csoneson,
Sorry for the oversight.
I have added my first name to Zenodo metadata and updated the version to `v1.3.2` to be consistent with the GitHub repo tag.
Thank you!
