### 2026-04-26 #9066 @stephwon
Hi @csoneson,

While reviewing the proof, we noticed that k-mer and k-size were not italicized, and there was a grammatical error at line 114 in the .md file, so we corrected these and rendered an updated proof [here](https://github.com/openjournals/joss-reviews/issues/9066#issuecomment-4322235161). Everything else is unchanged. Would it be possible to use the updated version as the final proof?

Thank you!
