### 2026-04-26 #9066 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
