### 2026-04-26 #9066 @stephwon
Hi @csoneson,
Thank you for your patience, and we apologize for the inconvenience. All errors have been addressed and double-checked; everything should now be in order. Please use the recently updated proof rendered [here](https://github.com/openjournals/joss-reviews/issues/9066#issuecomment-4323270434). We are ready to move forward with the AEiC final submission. Thank you!
